Shrink client bundle
