+++ src/auth/actions.ts
@@ -1,18 +1,20 @@
 'use server';
 
 import {
+  auth,
+  signIn,
+  signOut,
+} from '@/auth/server';
+import type { Session } from 'next-auth';
+import { redirect } from 'next/navigation';
+import {
+  generateAuthSecret,
   KEY_CALLBACK_URL,
   KEY_CREDENTIALS_CALLBACK_ROUTE_ERROR_URL,
   KEY_CREDENTIALS_SIGN_IN_ERROR,
   KEY_CREDENTIALS_SIGN_IN_ERROR_URL,
   KEY_CREDENTIALS_SUCCESS,
-  auth,
-  generateAuthSecret,
-  signIn,
-  signOut,
-} from '@/auth';
-import type { Session } from 'next-auth';
-import { redirect } from 'next/navigation';
+} from '.';
 
 export const signInAction = async (
   _prevState: string | undefined,
