+++ src/auth/cache.ts
@@ -1,4 +1,4 @@
 import { cache } from 'react';
-import { auth } from '@/auth';
+import { auth } from '@/auth/server';
 
 export const authCachedSafe = cache(() => auth().catch(() => null));
