+++ src/photo/form/index.ts
@@ -1,22 +1,13 @@
-import type { ExifData } from 'ts-exif-parser';
-import { DEFAULT_ASPECT_RATIO, Photo, PhotoDbInsert, PhotoExif } from '..';
+import { DEFAULT_ASPECT_RATIO, Photo, PhotoDbInsert } from '..';
 import {
-  convertTimestampToNaivePostgresString,
-  convertTimestampWithOffsetToPostgresString,
   generateLocalNaivePostgresString,
   generateLocalPostgresString,
   validationMessageNaivePostgresDateString,
   validationMessagePostgresDateString,
 } from '@/utility/date';
-import {
-  convertApertureValueToFNumber,
-  getAspectRatioFromExif,
-  getOffsetFromExif,
-} from '@/utility/exif';
 import { roundToNumber } from '@/utility/number';
 import { convertStringToArray, parameterize } from '@/utility/string';
 import { generateNanoid } from '@/utility/nanoid';
-import { GEO_PRIVACY_ENABLED } from '@/app/config';
 import { TAG_FAVS, getValidationMessageForTags } from '@/tag';
 import { MAKE_FUJIFILM } from '@/platforms/fujifilm';
 import { FujifilmRecipe } from '@/platforms/fujifilm/recipe';
@@ -271,46 +262,6 @@ export const convertPhotoToFormData = (photo: Photo): PhotoFormData => {
     favorite: photo.tags.includes(TAG_FAVS) ? 'true' : 'false',
   } as PhotoFormData);
 };
-
-// CREATE FORM DATA: FROM EXIF
-
-export const convertExifToFormData = (
-  data: ExifData,
-  film?: FujifilmSimulation,
-  recipeData?: FujifilmRecipe,
-): Omit<
-  Record<keyof PhotoExif, string | undefined>,
-  'takenAt' | 'takenAtNaive'
-> => ({
-  aspectRatio: getAspectRatioFromExif(data).toString(),
-  make: data.tags?.Make,
-  model: data.tags?.Model,
-  focalLength: data.tags?.FocalLength?.toString(),
-  focalLengthIn35MmFormat: data.tags?.FocalLengthIn35mmFormat?.toString(),
-  lensMake: data.tags?.LensMake,
-  lensModel: data.tags?.LensModel,
-  fNumber: (
-    data.tags?.FNumber?.toString() ||
-    convertApertureValueToFNumber(data.tags?.ApertureValue)
-  ),
-  iso: data.tags?.ISO?.toString() || data.tags?.ISOSpeed?.toString(),
-  exposureTime: data.tags?.ExposureTime?.toString(),
-  exposureCompensation: data.tags?.ExposureCompensation?.toString(),
-  latitude:
-    !GEO_PRIVACY_ENABLED ? data.tags?.GPSLatitude?.toString() : undefined,
-  longitude:
-    !GEO_PRIVACY_ENABLED ? data.tags?.GPSLongitude?.toString() : undefined,
-  film,
-  recipeData: JSON.stringify(recipeData),
-  ...data.tags?.DateTimeOriginal && {
-    takenAt: convertTimestampWithOffsetToPostgresString(
-      data.tags.DateTimeOriginal,
-      getOffsetFromExif(data),
-    ),
-    takenAtNaive:
-      convertTimestampToNaivePostgresString(data.tags.DateTimeOriginal),
-  },
-});
 
 // PREPARE FORM FOR DB INSERT
