+++ src/utility/html.ts
@@ -9,4 +9,4 @@ export const safelyParseFormattedHtml = (text: string) =>
 
 // Matches two or more <br> or <br /> tags in a row
 export const htmlHasBrParagraphBreaks = (text: string) =>
-  text.match(/(<br\s*\/?>){2}/i);
+  /(<br\s*\/?>){2}/i.test(text);
